Updating sbm.exe to use IHostedService so SIGTERM from containers is acknowledged Some logging updates
