+++ src/SqlSync.SqlBuild.UnitTest/SqlSync.SqlBuild.UnitTest.csproj
@@ -12,7 +12,7 @@
     <EmbeddedResource Include="Resources\sbx_package_tester.sbx" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Data.SqlClient" Version="3.0.0" />
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="3.0.1" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="5.0.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
     <PackageReference Include="MSTest.TestAdapter" Version="2.2.7" />
