+++ src/SqlSync.SqlBuild/SqlSync.SqlBuild.csproj
@@ -16,7 +16,7 @@
     <EmbeddedResource Include="Status\Xml\ServerReport_summary.xslt" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Data.SqlClient" Version="3.0.0" />
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="3.0.1" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="5.0.0" />
     <PackageReference Include="Microsoft.SqlServer.DacFx" Version="150.5164.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
